Implemented varying g
Implemented an ode solver that uses the hydrostatic equation to find the radius with pressure. To deal with unbound atmospheres, added a check to see whether the number of scale heights requires for the decrease in pressure gives a radius that's less than the Hill radius. This calculation assumes that the atmosphere is mostly isothermal and the Hill radius is calculated for a sun like star for now. If the radius is larger than the Hill radius, we revert to the constant g calculation and alert the user about the unboundedness of the atmosphere.