Updated Haze Model
Added two additional parameters: the scattering slope and the reference wavelength. The reference wavelength simply indicates the wavelength at which the Rayleigh scattering and (possibly) non-Rayleigh scattering have roughly the same scattering cross-section (the cross-section would only differ by some factor of 2 pi).
The scattering slope is a new parameter that can be fit in a retrieval. It accounts for non-Rayleight scattering in a simple way.

(Fixed small bugs in the test files that occured due to a change in the architecture)