Parametric form for radiative TP profile
This formulation is similar to the one found in Guillot 2010, Robinson & Caitling 2012, etc. It is regularly used in retrievals for emission spectra (e.g. Line et al 2013). By default, there's an optical channel (for absorbed light) and a thermal channel (for planet's thermal output). There's also the option of a second optical channel which offers additional flexibility in the TP profile.